Refactor the old initialization code to genesis.rs
